Fixed bug in destroy method of TorrentService

+++ src/main/java/com/github/yoep/popcorn/torrent/TorrentService.java
@@ -118,7 +118,7 @@ public class TorrentService {
     public void destroy() {
         Settings settings = settingsService.getSettings();
 
-        if (settings.isTorrentDirectoryCleaningEnabled()) {
+        if (settings.isTorrentDirectoryCleaningEnabled() && settings.getTorrentDirectory().exists()) {
             try {
                 log.info("Cleaning torrent directory {}", settings.getTorrentDirectory());
                 FileUtils.cleanDirectory(settings.getTorrentDirectory());
